refactor: #732 MainActivity, MainScreen 분리
