+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/xml/auth/login/LoginFragment.kt
@@ -16,7 +16,7 @@ import com.team.todoktodok.App
 import com.team.todoktodok.BuildConfig
 import com.team.todoktodok.R
 import com.team.todoktodok.databinding.FragmentLoginBinding
-import com.team.todoktodok.presentation.compose.discussion.DiscussionsActivity
+import com.team.todoktodok.presentation.compose.main.MainActivity
 import com.team.todoktodok.presentation.core.ExceptionMessageConverter
 import com.team.todoktodok.presentation.core.component.AlertSnackBar.Companion.AlertSnackBar
 import com.team.todoktodok.presentation.core.ext.repeatOnStarted
@@ -142,7 +142,7 @@ class LoginFragment : Fragment(R.layout.fragment_login) {
     }
 
     private fun moveToMain() {
-        val intent = DiscussionsActivity.Intent(requireContext())
+        val intent = MainActivity.Intent(requireContext())
         startActivity(intent)
         requireActivity().finish()
     }
