+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/xml/auth/signup/SignUpFragment.kt
@@ -11,7 +11,7 @@ import com.team.domain.model.member.NickNameException
 import com.team.todoktodok.App
 import com.team.todoktodok.R
 import com.team.todoktodok.databinding.FragmentSignupBinding
-import com.team.todoktodok.presentation.compose.discussion.DiscussionsActivity
+import com.team.todoktodok.presentation.compose.main.MainActivity
 import com.team.todoktodok.presentation.core.ExceptionMessageConverter
 import com.team.todoktodok.presentation.core.component.AlertSnackBar.Companion.AlertSnackBar
 import com.team.todoktodok.presentation.core.ext.repeatOnViewStarted
@@ -111,7 +111,7 @@ class SignUpFragment : Fragment(R.layout.fragment_signup) {
     }
 
     private fun moveToMain() {
-        val intent = DiscussionsActivity.Intent(requireContext())
+        val intent = MainActivity.Intent(requireContext())
         startActivity(intent)
         requireActivity().finish()
     }
